+++ CHANGELOG.md
@@ -11,9 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- BulkIngestRequest contains now one single list "items"
+- tests for ingestion adapted to BulkIngestRequest-Model
 ### Deprecated
 
 ### Removed
+- removed class  _BaseBulkIngestRequest for ingestion model
 
 ### Fixed